2020 Day 5

+++ AdventOfCode.Common/Template/DayX/DayX.cs
@@ -2,12 +2,12 @@ namespace AdventOfCode.Common.Template.DayX
 {
     public static class DayX
     {
-        public static int Part1()
+        public static long Part1()
         {
             return 0;
         }
 
-        public static int Part2()
+        public static long Part2()
         {
             return 0;
         }
